backend/requirements.txt
fastapi
uvicorn
requests
python-docx
reportlab
pydantic
